Import `chardet` in nodes-base
`iconv` does not have the capabilities to determine encodings. It would
make sense to import `chardet` into the codebase since it is is a small
and minimal package at only 22 KB, works in node/browser/native, works
on linux/mac/windows, has no dependencies, no native code or bindings,
has extensive code coverage, and is written in Typescript.

We will be using this in a later commit to fix the file corruptions.